Prevent external codecs from serializing types which are abstract, generated, or from the framework itself
