+++ src/Orleans.Analyzers/Orleans.Analyzers.csproj
@@ -37,4 +37,8 @@
     </EmbeddedResource>
   </ItemGroup>
 
+  <ItemGroup>
+    <AssemblyAttribute Remove="Orleans.Metadata.FrameworkPartAttribute"/>
+  </ItemGroup>
+
 </Project>
