+++ src/Orleans.Serialization.Abstractions/FrameworkPartAttribute.cs
@@ -1,0 +1,13 @@
+using System;
+using System.ComponentModel;
+
+namespace Orleans.Metadata;
+
+/// <summary>
+/// Specifies that an assembly does not contain application code.
+/// </summary>
+[AttributeUsage(AttributeTargets.Assembly)]
+[EditorBrowsable(EditorBrowsableState.Never)]
+public sealed class FrameworkPartAttribute : Attribute
+{
+}
